feat(#294): decompose frontend god files into focused modules
- api/scanner.ts (787 lines) → api/scanner/ with 7 focused sub-files
  (types, runs, results, configs, reviews, ws, misc) behind a facade index.ts
- Universe/quality domain extracted to api/universe.ts; 6 universe consumers
  updated to import directly from api/universe
- QualityReportModal.tsx (753 lines) → QualityReportModal/ directory with
  8 panel components, useQualityReport hook, and render tests per panel
- All 345 tests pass; tsc --noEmit clean; no file exceeds ~400 lines
